requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
boto3==1.34.0
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
requests